TabViewによる画面遷移を実装 アプリ全体の設定を管理するSettingを作成
