+++ PredicTalk/PredicTalk/PredicTalkApp.swift
@@ -12,6 +12,7 @@ struct PredicTalkApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .environmentObject(Setting())
         }
     }
 }
