+++ PredicTalk/PredicTalk/Sources/Setting.swift
@@ -1,0 +1,12 @@
+//
+//  Settings.swift
+//  PredicTalk
+//
+//  Created by 青原光 on 2023/11/14.
+//
+
+import Foundation
+
+class Setting: ObservableObject {
+    @Published var selectedLanguage: Language = .english
+}
